Added searchView - This allows the user to search for characters by account.
Switched order of setting online and death statuses. This allows
for seeing which characters have died when doing an account search and the user has multiple characters listed while being online.

Bug:
When an accountName cannot be found the json returned is the top 20 results of the league. Attempt to find a way to ignore this and throw an error if account isn't found.
Perhaps I can have a flag set for if its an account search. If it is and I see account fields in the json then I can throw that all away and throw an error instead.